update graph and requirements
