+++ requirements.txt
@@ -2,3 +2,6 @@ rdflib
 rdflib-jsonld
 uritemplate
 httplib2
+redis
+redisgraph
+hydrus
